add utilt, contains some function
