+++ .gitignore
@@ -22,3 +22,4 @@ docs/_build
 htmlcov/
 .xprocess
 .vscode
+.idea
