+++ src/flask_caching/__init__.py
@@ -16,28 +16,26 @@ import string
 import uuid
 import warnings
 from collections import OrderedDict
-from typing import Any
-from typing import Callable
-from typing import List
-from typing import Optional
-from typing import Tuple
-from typing import Union
 
-from flask import current_app
-from flask import Flask
-from flask import request
-from flask import url_for
+from typing import Any, Callable, List, Optional, Tuple, Union
+
+from flask import current_app, Flask, request, url_for
 from markupsafe import Markup
 from werkzeug.utils import import_string
 
 from flask_caching.backends.base import BaseCache
 from flask_caching.backends.simplecache import SimpleCache
+from flask_caching.utils import (
+    want_args, get_arg_names,
+    get_arg_default, get_id,
+    function_namespace,
+    make_template_fragment_key
+)
 
 __version__ = "1.10.1"
 
 logger = logging.getLogger(__name__)
 
-TEMPLATE_FRAGMENT_KEY_TEMPLATE = "_template_fragment_cache_%s%s"
 SUPPORTED_HASH_FUNCTIONS = [
     hashlib.sha1,
     hashlib.sha224,
@@ -46,110 +44,6 @@ SUPPORTED_HASH_FUNCTIONS = [
     hashlib.sha512,
     hashlib.md5,
 ]
-
-# Used to remove control characters and whitespace from cache keys.
-valid_chars = set(string.ascii_letters + string.digits + "_.")
-delchars = "".join(c for c in map(chr, range(256)) if c not in valid_chars)
-null_control = ({k: None for k in delchars},)
-
-
-def wants_args(f: Callable) -> bool:
-    """Check if the function wants any arguments"""
-
-    argspec = inspect.getfullargspec(f)
-
-    return bool(argspec.args or argspec.varargs or argspec.varkw)
-
-
-def get_arg_names(f: Callable) -> List[str]:
-    """Return arguments of function
-
-    :param f:
-    :return: String list of arguments
-    """
-    sig = inspect.signature(f)
-    return [
-        parameter.name
-        for parameter in sig.parameters.values()
-        if parameter.kind == parameter.POSITIONAL_OR_KEYWORD
-    ]
-
-
-def get_arg_default(f: Callable, position: int):
-    sig = inspect.signature(f)
-    arg = list(sig.parameters.values())[position]
-    arg_def = arg.default
-    return arg_def if arg_def != inspect.Parameter.empty else None
-
-
-def get_id(obj):
-    return getattr(obj, "__caching_id__", repr)(obj)
-
-
-def function_namespace(f, args=None):
-    """Attempts to returns unique namespace for function"""
-    m_args = get_arg_names(f)
-
-    instance_token = None
-
-    instance_self = getattr(f, "__self__", None)
-
-    if instance_self and not inspect.isclass(instance_self):
-        instance_token = get_id(f.__self__)
-    elif m_args and m_args[0] == "self" and args:
-        instance_token = get_id(args[0])
-
-    module = f.__module__
-
-    if m_args and m_args[0] == "cls" and not inspect.isclass(args[0]):
-        raise ValueError(
-            "When using `delete_memoized` on a "
-            "`@classmethod` you must provide the "
-            "class as the first argument"
-        )
-
-    if hasattr(f, "__qualname__"):
-        name = f.__qualname__
-    else:
-        klass = getattr(f, "__self__", None)
-
-        if klass and not inspect.isclass(klass):
-            klass = klass.__class__
-
-        if not klass:
-            klass = getattr(f, "im_class", None)
-
-        if not klass:
-            if m_args and args:
-                if m_args[0] == "self":
-                    klass = args[0].__class__
-                elif m_args[0] == "cls":
-                    klass = args[0]
-
-        if klass:
-            name = klass.__name__ + "." + f.__name__
-        else:
-            name = f.__name__
-
-    ns = ".".join((module, name))
-    ns = ns.translate(*null_control)
-
-    if instance_token:
-        ins = ".".join((module, name, instance_token))
-        ins = ins.translate(*null_control)
-    else:
-        ins = None
-
-    return ns, ins
-
-
-def make_template_fragment_key(fragment_name: str, vary_on: List[str] = None) -> str:
-    """Make a cache key for a specific fragment name."""
-    if vary_on:
-        fragment_name = "%s_" % fragment_name
-    else:
-        vary_on = []
-    return TEMPLATE_FRAGMENT_KEY_TEMPLATE % (fragment_name, "_".join(vary_on))
 
 
 class Cache:
